# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  domain.Wishlist:
    properties:
      created_at:
        type: string
      description:
        type: string
      event_date:
        type: string
      id:
        type: string
      name:
        type: string
      public_token:
        type: string
      user_id:
        type: string
    type: object
...
  dto.CreateWishlistRequest:
    properties:
      description:
        type: string
      event_date:
        type: string
      name:
        type: string
    type: object
...
  dto.UpdateWishlistRequest:
    properties:
      description:
        type: string
      event_date:
        type: string
      name:
        type: string
    type: object
  response.ErrorPayload:
    properties:
      error:
        type: string
    type: object
info:
  contact: {}
  description: 'REST API для вишлистов: авторизация, CRUD вишлистов/подарков, публичный
    просмотр и резервирование.'
  title: Wishlist Service API
  version: "1.0"
paths:
...
  /wishlists:
    get:
      description: Возвращает список вишлистов текущего пользователя.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Wishlist'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorPayload'
      security:
      - BearerAuth: []
      summary: List wishlists
      tags:
      - Wishlists
    post:
      consumes:
      - application/json
      description: Создает вишлист текущего пользователя.
      parameters:
      - description: Wishlist create payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateWishlistRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Wishlist'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorPayload'
      security:
      - BearerAuth: []
      summary: Create wishlist
      tags:
      - Wishlists
  /wishlists/{id}:
    delete:
      description: Удаляет вишлист текущего пользователя.
      parameters:
      - description: Wishlist ID (UUID)
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorPayload'
      security:
      - BearerAuth: []
      summary: Delete wishlist
      tags:
      - Wishlists
    get:
      description: Возвращает вишлист по ID, если он принадлежит текущему пользователю.
      parameters:
      - description: Wishlist ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Wishlist'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorPayload'
      security:
      - BearerAuth: []
      summary: Get wishlist by ID
      tags:
      - Wishlists
    patch:
      consumes:
      - application/json
      description: Частично обновляет вишлист текущего пользователя.
      parameters:
      - description: Wishlist ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Wishlist patch payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateWishlistRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Wishlist'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorPayload'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorPayload'
      security:
      - BearerAuth: []
      summary: Update wishlist
      tags:
      - Wishlists
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 'JWT access token with Bearer prefix. Example: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"